defined all functions in EventPolicy and AttendeePolicy and applied them in the appropriate controllers

+++ app/Http/Controllers/Api/EventController.php
@@ -19,7 +19,7 @@ class EventController extends Controller
 
     public function index()
     {
-        // Gate::authorize('viewAny', Event::class);
+        Gate::authorize('viewAny', Event::class);
         $query = $this->loadRelationships(Event::query());
 
         return EventResource::collection(
@@ -29,7 +29,7 @@ class EventController extends Controller
 
     public function store(Request $request)
     {
-        // Gate::authorize('create', Event::class);
+        Gate::authorize('create', Event::class);
         $event = Event::create([
             ...$request->validate([
                 'name' => 'required|string|max:255',
@@ -45,7 +45,7 @@ class EventController extends Controller
 
     public function show(Event $event)
     {
-        // Gate::authorize('view', $event);
+        Gate::authorize('view', $event);
         return new EventResource($this->loadRelationships($event));
     }
 
@@ -57,7 +57,7 @@ class EventController extends Controller
         //     abort(403, 'You are not authorized to update this event.');
         // }
 
-        Gate::authorize('update-event', $event);
+        Gate::authorize('update', $event);
 
         $event->update(
             $request->validate([
@@ -73,7 +73,7 @@ class EventController extends Controller
 
     public function destroy(Event $event)
     {
-        // Gate::authorize('delete', $event);
+        Gate::authorize('delete', $event);
 
         $event->delete();
